add strchr

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: fgeslin <fgeslin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/18 10:43:27 by fgeslin           #+#    #+#             */
-/*   Updated: 2022/10/18 10:58:05 by fgeslin          ###   ########.fr       */
+/*   Updated: 2022/10/18 12:04:17 by fgeslin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
