+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: fgeslin <fgeslin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/18 11:22:17 by fgeslin           #+#    #+#             */
-/*   Updated: 2022/10/18 11:40:17 by fgeslin          ###   ########.fr       */
+/*   Updated: 2022/10/18 12:04:36 by fgeslin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	int		i;
+	size_t	i;
 	char	*str;
 
 	if (!s)
